Refactor codegen module into submodules
- Extracted `utils.rs` for text manipulation (sanitize_name, transliterate).
- Extracted `types.rs` for type conversion (to_rust_type).
- Extracted `expressions.rs` for expression generation (generate_expr).
- Extracted `statements.rs` for statement generation (generate_statement).
- Updated `mod.rs` to re-export and orchestrate.
- Preserved existing logic and tests.
- Addressed warnings.

This change improves maintainability and cohesion by breaking the monolithic `codegen/mod.rs` file.
